requirements.txt
Django==4.2.7
django-bootstrap5==23.3
Pillow>=10.0.0
python-decouple==3.8
